Rename module encoding -> binary
