+++ packages/std/src/binary.rs
